+++ src/Core/Dfe.Complete.Application/Projects/Queries/ListAllProjects/ListAllProjectsForLocalAuthority.cs
@@ -13,7 +13,7 @@ public record ListAllProjectsForLocalAuthorityQuery(
     ProjectType? Type = null)
     : PaginatedRequest<PaginatedResult<List<ListAllProjectsResultModel>>>;
 
-public class ListAllProjectsForLocalAuthority(IListAllProjectsByFilterQueryService listAllProjectsByFilterQueryService)
+public class ListAllProjectsForLocalAuthority(IListAllProjectsQueryService listAllProjectsQueryService)
     : IRequestHandler<ListAllProjectsForLocalAuthorityQuery, PaginatedResult<List<ListAllProjectsResultModel>>>
 {
     public async Task<PaginatedResult<List<ListAllProjectsResultModel>>> Handle(
@@ -21,7 +21,7 @@ public class ListAllProjectsForLocalAuthority(IListAllProjectsByFilterQueryServi
     {
         try
         {
-            var projectsForLa = await listAllProjectsByFilterQueryService.ListAllProjectsByFilter(request.State, request.Type, localAuthorityCode: request.LocalAuthorityCode).ToListAsync(cancellationToken);
+            var projectsForLa = await listAllProjectsQueryService.ListAllProjects(request.State, request.Type, localAuthorityCode: request.LocalAuthorityCode).ToListAsync(cancellationToken);
             
             var paginatedResultModel = projectsForLa.Select(proj =>
                     ListAllProjectsResultModel.MapProjectAndEstablishmentToListAllProjectResultModel(