+++ src/Core/Dfe.Complete.Application/Projects/Queries/ListAllProjects/ListAllProjectsForTeam.cs
@@ -11,10 +11,11 @@ public record ListAllProjectsForTeamQuery(
     ProjectTeam Team,
     ProjectState? ProjectStatus,
     ProjectType? Type,
-    AssignedToState? AssignedToState = null)
+    AssignedToState? AssignedToState = null,
+    OrderProjectQueryBy? OrderBy = null)
     : PaginatedRequest<PaginatedResult<List<ListAllProjectsResultModel>>>;
 
-public class ListAllProjectsForTeamQueryHandler(IListAllProjectsByFilterQueryService listAllProjectsByFilterQueryService)
+public class ListAllProjectsForTeamQueryHandler(IListAllProjectsQueryService listAllProjectsQueryService)
     : IRequestHandler<ListAllProjectsForTeamQuery, PaginatedResult<List<ListAllProjectsResultModel>>>
 
 {
@@ -23,8 +24,8 @@ public class ListAllProjectsForTeamQueryHandler(IListAllProjectsByFilterQuerySer
     {
         try
         {
-            var projectsForTeam = await listAllProjectsByFilterQueryService.ListAllProjectsByFilter(
-                request.ProjectStatus, request.Type, request.AssignedToState, team: request.Team).ToListAsync(cancellationToken);
+            var projectsForTeam = await listAllProjectsQueryService.ListAllProjects(
+                request.ProjectStatus, request.Type, request.AssignedToState, team: request.Team, orderBy: request.OrderBy).ToListAsync(cancellationToken);
 
             var paginatedResultModel = projectsForTeam.Select(proj =>
                     ListAllProjectsResultModel.MapProjectAndEstablishmentToListAllProjectResultModel(proj.Project!, proj.Establishment))