+++ src/Frontend/Dfe.Complete/Extensions/DateTimeExtensions.cs
@@ -1,5 +1,6 @@
 using System.Globalization;
 using Dfe.Complete.Constants;
+using DfE.CoreLibs.Utilities.Constants;
 
 namespace Dfe.Complete.Extensions
 {
@@ -36,6 +37,9 @@ namespace Dfe.Complete.Extensions
 			return new DateTime(thisMonth.Year + yearsToAdd, month, 1);
 		}
 
+		public static string ToDateMonthYearString(this DateTime dateTime)
+			=> dateTime.ToString(DateFormatConstants.MonthAndYearFormat, GbCulture);
+
 		public static string ToDateMonthYearString(this DateTime? dateTime)
 		{
 			if (!dateTime.HasValue)
@@ -43,7 +47,18 @@ namespace Dfe.Complete.Extensions
 				return string.Empty;
 			}
 
-			return dateTime.Value.ToString(DateFormatConstants.MonthAndYearFormat, GbCulture);
+			return dateTime.Value.ToDateMonthYearString();
 		}
-    }
+		
+		public static string ToFullDateMonthYearString(this DateTime? dateTime)
+		{
+			if (!dateTime.HasValue)
+			{
+				return string.Empty;
+			}
+
+			return dateTime.Value.ToString(DateFormatConstants.FullMonthAndYearFormat, GbCulture);
+		}
+	}
+	
 }