+++ src/Frontend/Dfe.Complete/Pages/Projects/List/ProjectsByMonth/ConversionProjects.cshtml.cs
@@ -27,7 +27,7 @@ namespace Dfe.Complete.Pages.Projects.List.ProjectsByMonth
 
             var datetime = (DateTime?)date.ToDateTime(default);
             
-            DateString = datetime.ToDateMonthYearString();
+            DateString = datetime.ToFullDateMonthYearString();
             ViewData[TabNavigationModel.ViewDataKey] = AllProjectsTabNavigationModel;
             var listProjectQuery = new ListProjectsByMonthsQuery(date, null, ProjectState.Active, ProjectType.Conversion, PageNumber-1, PageSize);
             var response = await sender.Send(listProjectQuery);